+++ src/includes/classes/models.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * @package Abricos
+ * @subpackage Bos
+ * @copyright 2011-2015 Alexander Kuzmin
+ * @license http://opensource.org/licenses/mit-license.php MIT License (MIT)
+ * @author Alexander Kuzmin <roosit@abricos.org>
+ */
+
+/**
+ * Class BosSummary
+ */
+class BosSummary extends AbricosModel {
+    protected $_structModule = 'bos';
+    protected $_structName = 'Summary';
+}
+
+class BosSummaryList extends AbricosModelList {
+
+}
+
+?>