+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 - license changed to MIT
 - a completely refactoring source for core version 0.5.7
-
+- add online summary of all applications